refactor(store-core): extract checkpoint handlers into handlers/checkpoint.ts (audit P2-3)
Seventh slice of the P2-3 split. Move the checkpoint message handlers
(checkpoint_created / checkpoint_list / checkpoint_restored) and the
CheckpointRestoredPayload interface into a new handlers/checkpoint.ts.

Pure move, no logic change: the extracted block is byte-identical to the
removed lines (proven by diff). Imports Checkpoint (../types) and
resolveSessionId (./_shared). The Checkpoint type was used only by this
family, so it is dropped from the barrel's ../types import. The barrel
re-exports via `export * from './checkpoint'`, so the public surface is
unchanged.

handlers/index.ts: 4942 -> 4869 lines.
